Update to v1.4.
v1.4:
* Autodetection of Old 3DS/New 3DS EmuNANDs for both input/output
procedures, which removes the need of manually selecting a 3DS model.
* Added support for additional EmuNAND9 offsets. Every EmuNAND/RedNAND
is now created using the 'default' setup size (except if you're creating
additional EmuNANDs, in which case the same layout from EmuNAND #1 will
be applied).
* The program now shows the capacity of each drive displayed in the
drop-down lists. Furthermore, only drives with a capacity greater than 1
GiB will be displayed in the drop-down lists.
* The program now checks if a valid EmuNAND is available at the selected
slot before writing the name to the SD card.
* Fixed a bug that made the "Refresh" button also write the EmuNAND name
instead of just reading it.
* The FAT32 partition label is now properly written during the format
procedure.
* Removed the boot.bin modification feature.
* Fixed some more threading bugs.