Update to v1.7.
v1.7:
* Completely revamped the EmuNAND parsing code.
* Added compatibility with multiple NAND layouts in the same SD card.
The program is now able to detect, read and write EmuNANDs that don't
share the same layout as the first one.
* Every NAND inject operation now defaults to the minimum size possible,
including the format procedure. This should fix some compatibility
problems with Luma3DS.
* If a new NAND is injected and the size of the previous NAND was
greater than the minimum size possible, and if it wasn't stored as a
RedNAND, its NCSD header will be wiped from the SD card to avoid
problems.